scripts/demo_ecommerce/requirements.txt
faker==26.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
